infrastructure/aol-core/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
grpcio-health-checking==1.60.0
protobuf==4.25.1
python-consul==1.1.0
pyyaml==6.0.1
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
redis==5.0.1
aiohttp==3.9.1
tenacity==8.2.3

